+++ src/Methods/Umbrella.cpp
@@ -60,7 +60,7 @@ namespace SSAGES
 
 			// Update forces.
 			for(size_t j = 0; j < forces.size(); ++j)
-				for(size_t k = 0; k < forces[j].size(); ++k)
+				for(int k = 0; k < forces[j].size(); ++k)
 					forces[j][k] -= D*grad[j][k];
 		}
